Fixed bootstrap and update.sh ignoring .gitconfig
The grep filter I had in both bootstrap.sh and update.sh was set up to
ignore some git metadata files, but also ignored .gitconfig. Changed to
only ignore a few explicit git files.

The one potential downside to this fix is that if git adds another
metadata with the .git prefix, that file will need to be added to both
scripts.