Fixed zsh location on a per-OS basis
Before, bootstrap.sh assumed that zsh was located at /bin/zsh. However,
on FreeBSD, zsh is located at /usr/local/bin/zsh. bootstrap.sh will now
check to see which OS we're running and make sure we're looking for zsh
in the right place.

This also has the unintended effect of not changing the shell on
OSX-based systems. I don't have access to an OSX system at this point,
but if I do in the future I'll need to improve support for it.