Fixed bootstrap.sh choking on certain symlinks
Before this commit, backing up existing dotfiles, bootstrap.sh previous
choked if the target file was a symlink that pointed to an invalid or
deleted target.

Now, when the script backs up existing user dotfiles, it'll first test
to see if each file is a symlink and back up that symlink. Otherwise,
it'll back up each file as it normally would.