fix(alacritty): support emoji, etc. characters in title bar
Previously, any windows (e.g. a Firefox browser window, which displays
the current page's title in the title bar) that used an emoji in their
title bars would be displayed as a box character; the title bar did not
have enough room to display this character by default, so it would also
abruptly resize the title bar to accomodate the substitute character.

This changeset updates the Sway config to use the Noto Sans Mono font,
which has support for emojis. It also changes the default size of the
title bar to more comfortably accomodate characters like emojis.