feat(emacs): move emacs config to XDG config directory
As of emacs 27, storing the emacs configuration in an XDG-compatible
location (`$XDG_CONFIG_HOME/emacs`) is supported. This change moves the
emacs configuration to this location.

Note that if `~/.emacs` exists, emacs will prefer this configuration
over `$XDG_CONFIG_HOME/emacs`. You must remove the configuration at
`~/.emacs` for this to work. For more info, see:
https://github.com/emacs-mirror/emacs/commit/13995f31a219bfcb24da5887136583cbf2deff4c