fix(linux):
Unlike X11, under Wayland applications aren't permitted to access each
other's screen content; this is great for security, but breaks
screensharing. In order to enable screensharing, a program called
`xdg-desktop-portal` provides an API for accessing other applications'
screen content (and, optionally, entire screens' content).
`xdg-desktop-portal` launches a corresponding protocol-specific adapter
(e.g. `xdg-desktop-portal-gtk` for GNOME, `xdg-desktop-portal-wlr` for
`wlroots`-based DEs) depending on the value of `XDG_CURRENT_DESKTOP`.

Sway doesn't set this value and has rationale for not doing so
(https://github.com/swaywm/sway/pull/4876). This is a pretty hacky fix
because it sets it no matter which DE launches, but there really aren't
any good options for setting this (the ideal is probably to set it in
`sway.desktop`, but that must be maintained systemwide and is updated
any time the `sway` package is updated).